:sparkles: redis 테스트 수정
