+++ src/main/java/kr/hhplus/be/server/domain/token/repository/TokenRepository.java
@@ -6,8 +6,6 @@ public interface TokenRepository {
 
 	String createWaitToken(final String userId);
 
-	Integer countWaitToken();
-
 	List<String> getAllWaitTokens();
 
 	Long updateWaitTokens(final List<String> tokens);
